label-stick-be/service-gateway/pyproject.toml
[tool.poetry]
name = "service-gateway"
version = "0.1.0"
description = ""
authors = ["Phạm Tân Tị <kurokid0206@gmail.com>"]
readme = "README.md"
# packages = [{include = "service_gateway"}]

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = "^0.189.0"
uvicorn = "^0.22.0"
pydantic = "^1.10.9"
fastapi = "^0.98.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
